feat(storefront): add storefront version check and show banner correctly | hj
